requirements.txt
# Rectify Perspective - Dépendances
# Python 3.11+

PySide6>=6.6.0
opencv-python>=4.9.0
numpy>=1.24.0
Pillow>=10.0.0
piexif>=1.1.3
